chore(*): support old version of the plugin
