+++ .run/Run Plugin.run.xml
@@ -3,11 +3,6 @@
     <output_file path="$PROJECT_DIR$/../../console logs" />
     <log_file alias="idea.log" path="$PROJECT_DIR$/build/idea-sandbox/system/log/idea.log" />
     <ExternalSystemSettings>
-      <option name="env">
-        <map>
-          <entry key="DASHBOARD_URL" value="http://localhost:8080" />
-        </map>
-      </option>
       <option name="executionName" />
       <option name="externalProjectPath" value="$PROJECT_DIR$" />
       <option name="externalSystemIdString" value="GRADLE" />
